refactor: tui
