+++ cmd/cli/tui/controls/header/header_test.go
@@ -3,10 +3,7 @@ package header
 import (
 	"testing"
 
-	tea "github.com/charmbracelet/bubbletea"
 	"github.com/google/go-cmp/cmp"
-	"github.com/jonathanhope/armaria/cmd/cli/tui/msgs"
-	"github.com/jonathanhope/armaria/cmd/cli/tui/utils"
 )
 
 const Name = "header"
@@ -15,49 +12,49 @@ func TestCanUpdateWidth(t *testing.T) {
 	gotModel := HeaderModel{
 		name: Name,
 	}
-	gotModel, gotCmd := gotModel.Update(msgs.SizeMsg{Name: Name, Width: 1})
+	gotModel.Resize(1)
 
 	wantModel := HeaderModel{
 		name:  Name,
 		width: 1,
 	}
 
-	verifyUpdate(t, gotModel, wantModel, gotCmd)
+	verifyUpdate(t, gotModel, wantModel)
 }
 
 func TestCanUpdateNav(t *testing.T) {
 	gotModel := HeaderModel{}
-	gotModel, gotCmd := gotModel.Update(msgs.BreadcrumbsMsg("nav"))
+	gotModel.SetBreadcrumbs("breadcrumbs")
 
 	wantModel := HeaderModel{
-		nav: "nav",
+		breadcrumbs: "breadcrumbs",
 	}
 
-	verifyUpdate(t, gotModel, wantModel, gotCmd)
+	verifyUpdate(t, gotModel, wantModel)
 }
 
 func TestCanMarkBusy(t *testing.T) {
 	gotModel := HeaderModel{}
-	gotModel, gotCmd := gotModel.Update(msgs.BusyMsg{})
+	gotModel.SetBusy()
 
 	wantModel := HeaderModel{
 		busy: true,
 	}
 
-	verifyUpdate(t, gotModel, wantModel, gotCmd)
+	verifyUpdate(t, gotModel, wantModel)
 }
 
 func TestCanMarkFree(t *testing.T) {
 	gotModel := HeaderModel{
 		busy: true,
 	}
-	gotModel, gotCmd := gotModel.Update(msgs.FreeMsg{})
+	gotModel.SetFree()
 
 	wantModel := HeaderModel{
 		busy: false,
 	}
 
-	verifyUpdate(t, gotModel, wantModel, gotCmd)
+	verifyUpdate(t, gotModel, wantModel)
 }
 
 func TestBusy(t *testing.T) {
@@ -71,12 +68,10 @@ func TestBusy(t *testing.T) {
 	}
 }
 
-func verifyUpdate(t *testing.T, gotModel HeaderModel, wantModel HeaderModel, gotCmd tea.Cmd) {
+func verifyUpdate(t *testing.T, gotModel HeaderModel, wantModel HeaderModel) {
 	unexported := cmp.AllowUnexported(HeaderModel{})
 	modelDiff := cmp.Diff(gotModel, wantModel, unexported)
 	if modelDiff != "" {
 		t.Errorf("Expected and actual models different:\n%s", modelDiff)
 	}
-
-	utils.CompareCommands(t, gotCmd, nil)
 }
