+++ cmd/cli/tui/msgs/msgs.go
@@ -11,11 +11,12 @@ const (
 )
 
 // View is which View to show.
-type View string
+type View int
 
 const (
-	ViewBooks View = "books" // show a listing of books
-	ViewError View = "error" // show an error
+	ViewNone  View = iota // no view
+	ViewBooks             // show a listing of books
+	ViewError             // show an error
 )
 
 // ErrorMsg is a message that contains an error.
@@ -27,137 +28,27 @@ func (e ErrorMsg) Error() string { return e.Err.Error() }
 // ViewMsg is a message that contains which view to show.
 type ViewMsg View
 
-// SizeMsg is a message to inform the component it needs to resize.
-type SizeMsg struct {
-	Name   string // name of the target component
-	Width  int    // the max width the component can occupy
-	Height int    // the max height the component can occupy
-}
-
-// BusyMsg is a message that denotes the writer is busy.
-type BusyMsg struct{}
-
-// FreeMsg is a message that denotes the writer is free again.
-type FreeMsg struct{}
-
-// FolderMsg is a message that changes the current folder.
-type FolderMsg string
-
-// BreadcrumbsMsg is a message that changes the current breadcrumbs being displayed.
-type BreadcrumbsMsg string
-
-// FiltersMsg is published when the filters change.
-type FiltersMsg struct {
-	Name    string   // name of the target component
-	Filters []string // current set of filters
-}
-
-// InputModeMsg is used to change to input mode.
-type InputModeMsg struct {
-	Name      string // name of the target component
-	InputMode bool   // if true the footer will be in input mode
-	Prompt    string // the prompt to show
-	Text      string // the text to start the input with
-	MaxChars  int    // the maximum number of chars to allow
-}
-
-// InputCancelledMsg is published when the user cancels input.
-type InputCancelledMsg struct {
-	Name string // name of the component which is collecting input.
-}
-
-// InputConfirmedMsg is published when the user confirms input.
-type InputConfirmedMsg struct {
-	Name string // name of the component which is collecting input.
-}
-
-// InputChangedMsg is published when the value in an input changes.
-type InputChangedMsg struct {
-	Name string // name of the target component
-}
-
-// FocusMsg is used to focus an input.
-type FocusMsg struct {
-	Name     string // name of the target component
-	MaxChars int    // maximum number of chars to allow
-}
-
-// BlurMsg is published when the input blurs.
-type BlurMsg struct {
-	Name string // name of the target component
-}
-
-// TextMsg sets the text of an input.
-type TextMsg struct {
-	Name string // name of the target component
-	Text string // text to set
-}
-
-// PromptMsg sets the prompt of an input.
-type PromptMsg struct {
-	Name   string // name of the target component
-	Prompt string // prompt to use
-}
-
-// BlinkMsg is a message that causes the cursor to blink.
-type BlinkMsg struct {
-	Name string // name of the target component
-}
-
-// TypeaheadItem is the model for an item in the typeahead.
-type TypeaheadItem struct {
-	Value string // hidden identifier
-	Label string // visible text
-	New   bool   // if true this is a new item that didn't exist before
-}
-
-// UnfilteredQueryFn is a function that returns typeahead items when no filter is active.
-type UnfilteredQueryFn func() ([]TypeaheadItem, error)
-
-// FilteredQueryFn is a function that returns the typeahead items when a filter is active.
-type FilteredQueryFn func(query string) ([]TypeaheadItem, error)
-
-// TypeaheadModeMsg is used to change to typeahead mode.
-type TypeaheadModeMsg struct {
-	Name            string            // name of the target component
-	InputMode       bool              // if true the typeahead will be in input mode
-	Prompt          string            // the prompt to show
-	Text            string            // the text to start the input with
-	MaxChars        int               // the maximum number of chars to allow
-	UnfilteredQuery UnfilteredQueryFn // returns results when there isn't enough input
-	FilteredQuery   FilteredQueryFn   // returns results when there's enough input
-	MinFilterChars  int               // the minumum number of chars needed to filter
-	Operation       string            // the operation the typeahead is for
-	IncludeInput    bool              // if true include the current input as an option
-}
-
-// TypeaheadConfirmedMsg is published when an option is selected..
-type TypeaheadConfirmedMsg struct {
-	Name      string        // name of the target typeahead
-	Value     TypeaheadItem // the value that was selected
-	Operation string        // the operation the typeahead is for
-}
-
-// SelectionConfirmedMsg is published when an option selection is cancelled.
-type TypeaheadCancelledMsg struct {
-	Name string // name of the target typeahead
-}
-
-// ShowHelpMsg is used to bring up a help screen.
-type ShowHelpMsg struct {
-	Name string // name of the target help screen
-}
-
 // SelectionChangedMsg is published when the table selection changes.
 type SelectionChangedMsg[T any] struct {
-	Name      string // name of the target component
+	Name      string // name of the target control
 	Empty     bool   // whether the frame is empty
 	Selection T      // the selected item (if any) in the frame
 }
 
 // DataMsg is a message to update the data in the table.
 type DataMsg[T any] struct {
-	Name string    // name of the target component
+	Name string    // name of the target control
 	Data []T       //the data to show in the scrolltable
 	Move Direction // optionally adjust the cursor
 }
+
+// InputChangedMsg is published when the value in an input changes.
+type InputChangedMsg struct {
+	Name string // name of the target control
+}
+
+// BreadcrumbsMsg is a message that changes the current breadcrumbs being displayed.
+type BreadcrumbsMsg string
+
+// FolderMsg is a message that changes the current folder.
+type FolderMsg string
