+++ cmd/cli/tui/views/errorview/error.go
@@ -11,16 +11,13 @@ import (
 // model is the model for an error view.
 // The error view is used to display an error if one occurs.
 type model struct {
-	activeView msgs.View // which view is currently being shown
-	err        error     // the error that occurred
-	width      int       // the current width of the screen
+	err   error // the error that occurred
+	width int   // the current width of the screen
 }
 
 // InitialModel builds the model.
 func InitialModel() tea.Model {
-	return model{
-		activeView: msgs.ViewBooks,
-	}
+	return model{}
 }
 
 // Update handles a message.
@@ -30,9 +27,6 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.WindowSizeMsg:
 		m.width = msg.Width
 
-	case msgs.ViewMsg:
-		m.activeView = msgs.View(msg)
-
 	case msgs.ErrorMsg:
 		m.err = msg.Err
 		return m, func() tea.Msg { return msgs.ViewMsg(msgs.ViewError) }
@@ -40,9 +34,7 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 	case tea.KeyMsg:
 		switch msg.String() {
 		case "ctrl+c", "q":
-			if m.activeView == msgs.ViewError {
-				return m, tea.Quit
-			}
+			return m, tea.Quit
 		}
 	}
 
@@ -51,10 +43,6 @@ func (m model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
 
 // View renders the model.
 func (m model) View() string {
-	if m.activeView != msgs.ViewError {
-		return ""
-	}
-
 	return lipgloss.
 		NewStyle().
 		Width(m.width).
